+++ examples/workflow-theia/src/browser/workflow-glsp-client-contribution.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2019-2022 EclipseSource and others.
+ * Copyright (c) 2019-2023 EclipseSource and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -13,9 +13,10 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
-import { Args, MaybePromise } from '@eclipse-glsp/client';
-import { BaseGLSPClientContribution } from '@eclipse-glsp/theia-integration/lib/browser';
-import { injectable } from '@theia/core/shared/inversify';
+import { Args, ConnectionProvider, GLSPClient, MaybePromise } from '@eclipse-glsp/client';
+import { BaseGLSPClientContribution, TheiaJsonrpcGLSPClient } from '@eclipse-glsp/theia-integration/lib/browser';
+import { EnvVariablesServer } from '@theia/core/lib/common/env-variables';
+import { inject, injectable } from '@theia/core/shared/inversify';
 import { WorkflowLanguage } from '../common/workflow-language';
 
 export interface WorkflowInitializeOptions {
@@ -25,6 +26,9 @@ export interface WorkflowInitializeOptions {
 
 @injectable()
 export class WorkflowGLSPClientContribution extends BaseGLSPClientContribution {
+    @inject(EnvVariablesServer)
+    protected readonly envVariablesServer: EnvVariablesServer;
+
     readonly id = WorkflowLanguage.contributionId;
     readonly fileExtensions = WorkflowLanguage.fileExtensions;
 
@@ -33,5 +37,27 @@ export class WorkflowGLSPClientContribution extends BaseGLSPClientContribution {
             ['timestamp']: new Date().toString(),
             ['message']: 'Custom Options Available'
         };
+    }
+
+    protected override async createGLSPClient(connectionProvider: ConnectionProvider): Promise<GLSPClient> {
+        const webSocketPort = await this.getWebSocketPortFromEnv();
+        return new TheiaJsonrpcGLSPClient({
+            id: this.id,
+            connectionProvider,
+            messageService: this.messageService,
+            webSocketPort
+        });
+    }
+
+    protected async getWebSocketPortFromEnv(): Promise<number | undefined> {
+        const envVar = await this.envVariablesServer.getValue('WEBSOCKET_PORT');
+        if (envVar && envVar.value) {
+            const webSocketPort = Number.parseInt(envVar.value, 10);
+            if (isNaN(webSocketPort) || webSocketPort < 0 || webSocketPort > 65535) {
+                throw new Error('Value of environment variable WEBSOCKET_PORT is not a valid port');
+            }
+            return webSocketPort;
+        }
+        return undefined;
     }
 }